+++ src/modules/common/components/controlled-fields/radio-group/RadioGroup.tsx
@@ -2,8 +2,8 @@ import {
   Box,
   FormControlLabel,
   FormHelperText,
-  Radio,
   RadioGroup as MuiRadioGroup,
+  Radio,
 } from '@mui/material';
 import { useController } from 'react-hook-form';
